+++ Source/Sitecore.FakeDb.Tests/DbConcurrencyTest.cs
@@ -8,32 +8,55 @@
   using Sitecore.Xml;
   using Sitecore.Configuration;
   using Sitecore.Pipelines;
+  using Sitecore.Data;
 
   public class DbConcurrencyTest
   {
     [Fact]
     public void ShouldBeThreadLocalDb()
     {
-      // arrange
       var t1 = Task.Factory.StartNew(() =>
         {
-          var random = new Random();
           using (var db = new Db { new DbItem("Home") })
           {
-            Thread.Sleep(random.Next(1000));
+            Thread.Sleep(1000);
             db.GetItem("/sitecore/content/home").Should().NotBeNull("the Home item is expected");
           }
         });
 
       var t2 = Task.Factory.StartNew(() =>
         {
-          var random = new Random();
           using (var db = new Db())
           {
-            Thread.Sleep(random.Next(1000));
             db.GetItem("/sitecore/content/home").Should().BeNull("the Home item is not expected");
           }
         });
+
+      t1.Wait();
+      t2.Wait();
+    }
+
+    [Fact]
+    public void ShouldBeThreadLocalDataProvider()
+    {
+      var id = ID.NewID;
+
+      var t1 = Task.Factory.StartNew(() =>
+      {
+        using (var db = new Db { new DbTemplate(id) })
+        {
+          Thread.Sleep(1000);
+          db.Database.Templates[id].Should().NotBeNull("the Home template is expected");
+        }
+      });
+
+      var t2 = Task.Factory.StartNew(() =>
+      {
+        using (var db = new Db())
+        {
+          db.Database.Templates[id].Should().BeNull("the Home template is not expected");
+        }
+      });
 
       t1.Wait();
       t2.Wait();